requirements.txt
# requirements.txt
PyMuPDF==1.23.14
python-docx==1.1.0
python-pptx==0.6.23
pandas==2.1.4
openpyxl==3.1.2
tabula-py==2.9.0
Pillow==10.2.0